requirements.txt
streamlit
google-generativeai
python-dotenv
Image
pathlib
textwrap